requirements.txt
torch==2.1.0+cu118
torchvision==0.16.0+cu118
pyautogui==0.9.54
keyboard==0.13.5
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.23.5
pytesseract==0.3.13
tqdm==4.63.0